+++ R/load-read_glatos_detections.r
@@ -1,45 +1,43 @@
-#' @title 
-#' Read data from a GLATOS detection file
-#' 
+#' @title Read data from a GLATOS detection file
+#'
 #' @description Read data from a standard GLATOS detection (csv) file and return
-#' a data.frame of class \code{glatos_detections}.
+#'   a data.frame of class `glatos_detections`.
 #'
 #' @param det_file A character string with path and name of detection file in
 #'   standard GLATOS format (*.csv). If only file name is given, then the file
 #'   must be located in the working directory. File must be a standard GLATOS
-#'   file (e.g., \emph{xxxxx_detectionsWithLocs_yyyymmdd_hhmmss.csv}) submitted
-#'   via GLATOSWeb Data Portal \url{http://glatos.glos.us}.
-#'  
+#'   file (e.g., *xxxxx_detectionsWithLocs_yyyymmdd_hhmmss.csv*) submitted via
+#'   GLATOSWeb Data Portal <https://glatos.glos.us>.
+#'
 #' @param version An optional character string with the glatos file version
 #'   number. If NULL (default value) then version will be determined by
-#'   evaluating file structure. The only allowed values currently are
-#'   \code{NULL} and \code{"1.3"}. Any other values will trigger an error.
-#'  
-#' @details
-#' Data are loaded using \link[data.table]{fread} and timestamps are coerced to
-#' POSIXct using \link[fasttime]{fastPOSIXct}. All times must be in UTC timezone
-#' per GLATOS standard.
-#' 
-#' @details Column \code{animal_id} is considered a required column by many
-#'   other functions in this package, so it will be created if any records are
-#'   \code{NULL}. When created, it will be constructed from
-#'   \code{transmitter_codespace} and \code{transmitter_id}, separated by '-'.
-#' 
-#' @return A data.frame of class \code{glatos_detections}.
+#'   evaluating file structure. The only allowed values currently are `NULL` and
+#'   `"1.3"`. Any other values will trigger an error.
 #'
-#' @author C. Holbrook \email{cholbrook@usgs.gov}
+#' @details Data are loaded using [fread][data.table::fread] and timestamps are
+#' coerced to POSIXct using [fastPOSIXct][fasttime::fastPOSIXct]. All times must
+#' be in UTC timezone per GLATOS standard.
+#'
+#' @details Column `animal_id` is considered a required column by many other
+#'   functions in this package, so it will be created if any records are `NULL`.
+#'   When created, it will be constructed from `transmitter_codespace` and
+#'   `transmitter_id`, separated by '-'.
+#'
+#' @return A data.frame of class `glatos_detections`.
+#'
+#' @author C. Holbrook \email{cholbrook@@usgs.gov}
 #'
 #' @examples
 #' #get path to example detection file
 #' det_file <- system.file("extdata", "walleye_detections.csv",
 #'                          package = "glatos")
-#'                          
+#'
 #' #note that code above is needed to find the example file
 #' #for real glatos data, use something like below
-#' #det_file <- "c:/path_to_file/HECWL_detectionsWithLocs_20150321_132242.csv"           
-#'                          
+#' #det_file <- "c:/path_to_file/HECWL_detectionsWithLocs_20150321_132242.csv"
+#'
 #' det <- read_glatos_detections(det_file)
-#' 
+#'
 #' @importFrom lubridate parse_date_time
 #'
 #' @export
@@ -54,7 +52,10 @@ read_glatos_detections <- function(det_file, version = NULL) {
   #Identify detection file version
   id_det_version <- function(det_file){
     det_col_names <- names(data.table::fread(det_file, nrows = 0))
-    if(all(glatos:::glatos_detection_schema$v1.3$name %in% det_col_names)) { 
+    if(all(glatos:::glatos_detection_schema$v1.4$name %in% det_col_names)) { 
+      return("1.4") 
+    } else 
+      if(all(glatos:::glatos_detection_schema$v1.3$name %in% det_col_names)) { 
       return("1.3") 
     } else {
       stop("Detection file version could not be identified.")
@@ -68,10 +69,12 @@ read_glatos_detections <- function(det_file, version = NULL) {
     stop(paste0("Detection file version ", version," is not supported."))
   }
  
-  #-Detections v1.3----------------------------------------------------------------  
-  if (version == "1.3") {
+  #-Detections v1.3 or v1.4-----------------------------------------------------  
+  if(version %in% c("1.3", "1.4")) {
 
-    col_classes <- glatos:::glatos_detection_schema[["v1.3"]]$type
+    vversion <- paste0("v", version)
+    
+    col_classes <- glatos:::glatos_detection_schema[[vversion]]$type
     timestamp_cols <- which(col_classes == "POSIXct")
     date_cols <- which(col_classes == "Date")
     col_classes[c(timestamp_cols, date_cols)] <- "character"
@@ -84,22 +87,22 @@ read_glatos_detections <- function(det_file, version = NULL) {
     #  timestamp must be in UTC; and tz argument sets the tzone attr only
     options(lubridate.fasttime = TRUE)
     for (j in timestamp_cols) data.table::set(dtc, 
-                                  j = glatos:::glatos_detection_schema[["v1.3"]]$name[j], 
+                                  j = glatos:::glatos_detection_schema[[vversion]]$name[j], 
                       value = lubridate::parse_date_time(
-                           dtc[[glatos:::glatos_detection_schema[["v1.3"]]$name[j]]], 
+                           dtc[[glatos:::glatos_detection_schema[[vversion]]$name[j]]], 
                            orders="ymd HMS",
                                 tz = "UTC"))
     #coerce dates to date
     for (j in date_cols) {
-      data.table::set(dtc, j = glatos:::glatos_detection_schema[["v1.3"]]$name[j], 
-        value = ifelse(dtc[[glatos:::glatos_detection_schema[["v1.3"]]$name[j]]] == "", 
+      data.table::set(dtc, j = glatos:::glatos_detection_schema[[vversion]]$name[j], 
+        value = ifelse(dtc[[glatos:::glatos_detection_schema[[vversion]]$name[j]]] == "", 
                        NA, 
-                       dtc[[glatos:::glatos_detection_schema[["v1.3"]]$name[j]]]))
-      data.table::set(dtc, j = glatos:::glatos_detection_schema[["v1.3"]]$name[j], 
-        value = as.Date(dtc[[glatos:::glatos_detection_schema[["v1.3"]]$name[j]]]))
+                       dtc[[glatos:::glatos_detection_schema[[vversion]]$name[j]]]))
+      data.table::set(dtc, j = glatos:::glatos_detection_schema[[vversion]]$name[j], 
+        value = as.Date(dtc[[glatos:::glatos_detection_schema[[vversion]]$name[j]]]))
     }
   }
-  #-end v1.3----------------------------------------------------------------
+  #-end v1.3 or v1.4------------------------------------------------------------
 
   #create animal_id if missing
   anid_na <- is.na(dtc$animal_id)